+++ Tests/SentryTests/SentryClientTests.swift
@@ -798,10 +798,18 @@ class SentryClientTest: XCTestCase {
         XCTAssertEqual(1, exceptions.count)
         let exception = exceptions[0]
         XCTAssertEqual(error.domain, exception.type)
-        XCTAssertEqual("\(error.code)", exception.value)
+        
+        XCTAssertEqual("Code: \(error.code)", exception.value)
+        
         XCTAssertNil(exception.thread)
         
-        XCTAssertEqual([error.domain, String(error.code)], event.fingerprint)
+        guard let mechanism = exception.mechanism else {
+            XCTFail("Exception doesn't contain a mechanism"); return
+        }
+        XCTAssertEqual("NSError", mechanism.type)
+        XCTAssertNotNil(mechanism.error)
+        XCTAssertEqual(error.domain, mechanism.error?.domain)
+        XCTAssertEqual(error.code, mechanism.error?.code)
         
         assertValidDebugMeta(actual: event.debugMeta)
         assertValidThreads(actual: event.threads)