+++ CHANGELOG.md
@@ -14,7 +14,7 @@ you can take a look at
 [SDK fingerprinting](https://docs.sentry.io/platforms/apple/data-management/event-grouping/sdk-fingerprinting/)
 to group by domain only.
 
-
+- fix: Use mechanism meta for error grouping #946
 - fix: Sanitize SentryMechanism.data on serialize #947
 - feat: Add error to SentryEvent #944
 - fix: Mark SentryEvent.message as Nullable #943